partly catching up current trunk ...

+++ branches/user/grundl/examples/hydraulics/rigid_node/system.cc
@@ -34,7 +34,7 @@ System::System(const string &name, bool setvalued) : Group(name) {
   l23->setDirection("[0;0;0]");
 
   ConstrainedNode * n1 = new ConstrainedNode("n1");
-  n1->setpFunction(new ConstantFunction<double>(5e5));
+  n1->setpFunction(new ConstantFunction<double(double)>(5e5));
   addLink(n1);
   n1->addOutFlow(l12);
